+++ tests/testthat/test_pipeop_impute.R
@@ -401,7 +401,6 @@ test_that("More tests for Integers", {
     expect_false(any(is.na(result$data()$x)), info = po$id)
     expect_equal(result$missings(), c(t = 0, x = 0), info = po$id)
   }
-
 })
 
 test_that("impute, test rows and affect_columns", {